+++ Cargo.toml
@@ -22,3 +22,4 @@ git-version = "0.3"
 [dev-dependencies]
 proptest = "1.7"
 wiremock = "0.6"
+serial_test = "2"
